+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Enhancements
 
 - Added a compact JSON serialiser.
+- Added a compact JSON deserialiser.
 
 ## 0.2.0 (2016-04-29)